+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Fix `Slack\Objects\BlockElements\SelectMenu\StaticSelect`
   Contruct now takes a single option as `initial_option`.
 * `Slack\Objects\Blocks\ActionBlock` can now take an array of elements in contructor.
+* Fix `Slack\Http\Controllers\InteractionController` block_suggestions.
+* Fix missing user authentication in `Slack\Http\Middleware\SlackCommand`.
 
 ## [1.1.0] - 2020-11-10